move getStats into their own store
